condensed report
this contains a condensed report for work up to 7 May 2025. We categorise the inpatient mitigators into the three types of interest, perform the sampling and derivation of a single table based on the hierarchy. We also derive the counts for outpatient (which, if mitigatable, are de-adoption) and ED (redirection and substitution)