remove outpatient convert to tele from de-adoption cases
The group of strategies for converting outpatient attendances to tele-attendances is being flagged as de-adoption due to the method used. (This is just checking if the `attendkey` exists in `nhp.raw_data.opa_mitigators`, based on an implicit assumption that all outpatient mitigators were de-adoption). This is not the case as they are more about redirecting. For the purposes of this analysis, we will just group them together with "none" given they are efficiencies and we are just subsetting activity avoidance.